did most of image upload
did most of the image upload. Currently, it is able to know the default image and post it for register. It for some reason is to post the image itself in profile, but is giving a link. The link uses amazon's API to store it. I did also store a copy locally just in case someone wanted access to it. I am having 2 problems currently 1) the image not displaying on profile (even though it for the most part is for the register page. 2) a few problems with uploading new images rather than retrieving ones.